replace rusoto_s3 in columnq::table::delta
