+++ columnq/src/io/object_store.rs
@@ -22,6 +22,28 @@ pub async fn partition_key_to_reader(
     Ok(std::io::Cursor::new(bytes.to_vec()))
 }
 
+pub async fn partitions_from_path_iterator<'a, F, T, I>(
+    path_iter: I,
+    mut partition_reader: F,
+) -> Result<Vec<T>, ColumnQError>
+where
+    I: Iterator<Item = &'a str>,
+    F: FnMut(std::io::Cursor<Vec<u8>>) -> Result<T, ColumnQError>,
+{
+    let object_store_provider = ColumnQObjectStoreProvider {};
+    let mut partitions = vec![];
+
+    for path_str in path_iter {
+        let url = &Url::from_str(path_str).unwrap();
+        let client = object_store_provider.get_by_url(url)?;
+        let path = object_store::path::Path::from(&url.path()[1..]);
+        let reader = partition_key_to_reader(client.clone(), &path).await?;
+        partitions.push(partition_reader(reader)?);
+    }
+
+    Ok(partitions)
+}
+
 pub async fn partitions_from_uri<'a, F, T>(
     t: &'a TableSource,
     _uri: URIReference<'a>,
